Extract swagger setup
